Bug in VariableFile (probably legacy from a non finalized test)
not sure why, I defined variables_["ALPHA_TW_MIN"] = 30 and
variables_["ALPHA_TW_MAX"] = 30. This is possibly left over from a test.
Modified the Variable::operator[] to return references, in order to allow for
an expression such as variables_["ALPHA_TW_MIN"] = 30.
Note that prior to this change the assignement did not affect the variable in the
set but a tmp copy of the variable, that is the reason why this pb did not affect.
I am not sure however of other places where the code would convert degs to rads,
if this happens and if it has ever worked