Improved the plotDelta_ResidRes_Heel method visualizing data for several PHI_
The formulation of the plotDelta_ResidRes_Heel starts -- in the DYHS -- from Fn=
0.25. I have added some control points @Fn=0, 0.1, 0.2 in order to assure Res=0
for Fn<0.25. Unfortunately the spline interpolation oscillates for values Fn<0.25
and we have negative resistances -proportional to the heel angle- for Fn =0-.25

I conjecture this is the responsible for the phenomenon I have observed, of a
slight increase in velocity @negative heel angles.

Test this with:
plotDelta_ResidRes_Heel
twv, twa = 0 0
crew     = 0
flat     = 1
Fn_min   = .18
Fn_max   = .3