Attempt improving the polar plots.
Unfortunately I cannot correctly trigger Axes equal, because of some interaction
with the multiplot widget, I am afraid. So in order to get axes equal one must
trigger a mouseclickevent, or a mousemoveevent. Really not ideal.
This also affects the ranges, which are computed based on the current -non axes
equal- pixeled size of the plot.
In the end, one will get a multiplot widget with non axes equal plots. When
clicking, one sees the axes equal but the y-range going out of the range, thus
requiring scaling