This does not work still, though it goes toward the right direction...
Because the children of the SettingsItemBase are not deep copied, thus we are in
a situation where the children of the model and the reference model are in fact
the same objects.
I believe the key is in the copy ctor of SettingsItemBase, where I shallow copy
the children with : children_ = rhs.children_
because children_ is a QList<SettingsItemBase*>

I need to make a deep copy, but my initial tests were not conclusive.