Attempt to sync the variabletree with the settings - not entirely functional
Actually the variableFileParser is in sync with the reference variable model.
Here we try a sync with the current variableModel while this is being edited.
This is not possible as it would break the 'undo' mechanism arising while hitting
the 'cancel' button (restore the reference model to the current model).

It does not make sense to update the variableTreeModel while editing the reference
settings model. What we need is to update the variableTreeModel after assigning
the current settings mdoel to the reference settings model.