Corrected the update of the variableTree
The TreeTab is instantiated by the VPPSettingsDialog, which is its parent.
However, after construction it is assigned to a QTabWidget which lives in a
layout of the VPPSettingsDialog.
The Resulting hierarchy is therefore :
 MainWindow
    VPPSettingsDialog
       Layout
          QTabWidget
             TreeTab

The only relieable way to retrieve the mainWindow is to loop on this hierarchy
and search for a mainWindow.
I was so confused beacuse I did not expect the hierarchy to change after the
construction of the TreeTab (which is initially child of the VPPSettingsDialog).
But in fact the hierarchy change does make sense...

Questions:
Is this the best way to retrieve the MainWindow? Is it a good idea to retrieve
the MainWindow at all..?