UPDATED TEST BASELINE FOR THE FULLIPOPTRUNTEST
the difference was so small that it was undetectable. What has changed? Possibly
the macos system and perhaps the compiler version. I have tried some back and
forth between version without being able to determine the cause of the (small)
difference. I decide not to become crazy and accept the changes