Patch the scons qt5 tool and remove the repository internal to site_scons
The patch is in site_scons/site_tools/qt5/__init__.py where I modify the path to
the qt bin folder, that was pre-pended by "clang_64" that I do not want (at least
in this hard-coded fashion).

Also, site_scons is embedded into the build system - by now. While one can question
the suitability of such an option, we do not want a subrepo in the repo, or the
files won't be tracked.